Added UI-O-Matic
I haven't used the latest version of UI-O-Matic, but the old one was of great use.

You could argue this package should be used as a developer tool, but at GrowCreate we have given a client access to it so they can manage their database easily. This is why I have placed it under 'Backoffice extensions'.